Updated model Added components to dataset Changed DECIMAL types to float

+++ ensysmod/model/energy_storage.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, ForeignKey, DECIMAL
+from sqlalchemy import Column, Integer, ForeignKey, Float
 from sqlalchemy.orm import relationship
 
 from ensysmod.database.base_class import Base
@@ -13,14 +13,14 @@ class EnergyStorage(Base):
     ref_component = Column(Integer, ForeignKey("energy_component.id"), index=True, nullable=False, primary_key=True)
     ref_commodity = Column(Integer, ForeignKey("energy_commodity.id"), index=True, nullable=False)
 
-    charge_efficiency = Column(DECIMAL, nullable=True)
-    discharge_efficiency = Column(DECIMAL, nullable=True)
-    self_discharge = Column(DECIMAL, nullable=True)
+    charge_efficiency = Column(Float, nullable=True)
+    discharge_efficiency = Column(Float, nullable=True)
+    self_discharge = Column(Float, nullable=True)
     cyclic_lifetime = Column(Integer, nullable=True)
-    charge_rate = Column(DECIMAL, nullable=True)
-    discharge_rate = Column(DECIMAL, nullable=True)
-    state_of_charge_min = Column(DECIMAL, nullable=True)
-    state_of_charge_max = Column(DECIMAL, nullable=True)
+    charge_rate = Column(Float, nullable=True)
+    discharge_rate = Column(Float, nullable=True)
+    state_of_charge_min = Column(Float, nullable=True)
+    state_of_charge_max = Column(Float, nullable=True)
 
     # Relationships
     component = relationship("EnergyComponent")
